+++ react/components/pages/AddressCreate.tsx
@@ -1,17 +1,21 @@
 import React, { Component } from 'react'
-import { graphql } from 'react-apollo'
-import { compose, branch, renderComponent, withProps } from 'recompose'
+import { graphql, MutationResult, DataValue } from 'react-apollo'
+import { compose, branch, renderComponent } from 'recompose'
 import queryString from 'query-string'
-import { withRuntimeContext } from 'vtex.render-runtime'
+import { withRouter, RouteComponentProps } from 'vtex.my-account-commons/Router'
 
 import { withContentWrapper } from '../shared/withContentWrapper'
 import AddressCreateLoading from '../loaders/AddressCreateLoading'
 import AddressForm from '../Addresses/AddressForm'
 import ContentBox from '../shared/ContentBox'
-import getEmptyAddress from '../Addresses/emptyAddress'
-import CREATE_ADDRESS from '../../graphql/createAddress.gql'
-import GET_NEW_ADDRESS_DATA from '../../graphql/getNewAddressData.gql'
 import styles from '../../styles.css'
+import SAVE_ADDRESS, {
+  Args,
+  Result as SaveAddressResult,
+} from '../../graphql/saveAddress.gql'
+import CREATE_ADDRESS_INFO, {
+  Result as CustomerInfo,
+} from '../../graphql/customerGreeting.gql'
 
 export const headerConfig = {
   namespace: `${styles.addressCreate}`,
@@ -22,33 +26,45 @@ export const headerConfig = {
   },
 }
 
-class AddressCreate extends Component<Props, State> {
+function buildReceiverName(profile: CustomerInfo['profile']): string {
+  let lastName = ''
+  if (profile.lastName) {
+    lastName += ` ${profile.lastName}`
+  }
+
+  return `${profile.firstName ?? ''}${lastName}`
+}
+
+class AddressCreate extends Component<Props> {
   public state = {
     isLoading: false,
   }
 
-  private handleGoBack = () => {
-    const { history } = this.props
-
-    const parsed = queryString.parse(history.location.search)
-
-    history.push(
-      parsed.returnUrl ? parsed.returnUrl : '/addresses?success=true'
-    )
-  }
-
   private handleCreate = (address: Address) => {
-    const { createAddress, handleError } = this.props
-    const { addressId, addressQuery, ...addressFields } = address
+    const { saveAddress, handleError, history, location } = this.props
+    const { addressId, addressQuery, receiverName, ...addressFields } = address
 
     this.setState({ isLoading: true })
 
-    createAddress({
+    saveAddress({
       variables: {
-        addressFields: addressFields as AddressInput,
+        address: { ...addressFields } as AddressInput,
       },
     })
-      .then(() => this.handleGoBack())
+      .then(({ data }) => {
+        const parsed = queryString.parse(location.search)
+
+        if (parsed.returnUrl && data) {
+          history.push({
+            pathname: parsed.returnUrl,
+            search: queryString.stringify({
+              newAddressId: data.saveAddress.id,
+            }),
+          })
+        } else {
+          history.goBack()
+        }
+      })
       .catch(() => {
         handleError()
         this.setState({ isLoading: false })
@@ -57,21 +73,15 @@ class AddressCreate extends Component<Props, State> {
 
   public render() {
     const { isLoading } = this.state
-    const { runtime, shipsTo, profile } = this.props
-
-    const country =
-      shipsTo && shipsTo.length > 0 ? shipsTo[0] : runtime.culture.country
-
-    const address = getEmptyAddress(country, profile)
+    const { receiverName } = this.props
 
     return (
       <ContentBox shouldAllowGrowing maxWidthStep={6}>
         <AddressForm
-          address={address}
+          receiverName={receiverName}
           isLoading={isLoading}
           submitLabelId="vtex.store-messages@0.x::addresses.addAddress"
           onSubmit={this.handleCreate}
-          shipsTo={shipsTo}
           onError={this.props.handleError}
         />
       </ContentBox>
@@ -79,43 +89,37 @@ class AddressCreate extends Component<Props, State> {
   }
 }
 
-interface Data {
-  profile: Profile
-  logistics: {
-    shipsTo: string[]
+interface MappedResult {
+  loading: boolean
+  receiverName: string
+  data: DataValue<unknown>
+}
+
+type Props = MappedResult &
+  RouteComponentProps &
+  InjectedContentWrapperProps & {
+    saveAddress: (
+      args: Variables<Args>
+    ) => Promise<MutationResult<SaveAddressResult>>
   }
-}
-
-interface Profile {
-  firstName: string
-  lastName: string
-}
-
-interface Props extends InjectedContentWrapperProps {
-  profile: Profile
-  shipsTo: string[]
-  history: any
-  runtime: Runtime
-  createAddress: (args: Variables<CreateAddressArgs>) => Promise<void>
-}
-
-interface State {
-  isLoading: boolean
-}
 
 const enhance = compose<Props, void>(
-  graphql(GET_NEW_ADDRESS_DATA),
-  graphql(CREATE_ADDRESS, { name: 'createAddress' }),
+  graphql<{}, CustomerInfo, {}, MappedResult>(CREATE_ADDRESS_INFO, {
+    props: props => ({
+      loading: props.data ? props.data.loading : true,
+      data: props.data as DataValue<unknown>,
+      receiverName: props.data?.profile
+        ? buildReceiverName(props.data?.profile)
+        : '',
+    }),
+  }),
   branch(
-    ({ data }: { data: Data }) => data.profile == null,
+    ({ loading }: Props) => loading,
     renderComponent(AddressCreateLoading)
   ),
-  withProps(({ data }: { data: Data }) => ({
-    profile: data.profile,
-    shipsTo: data.logistics.shipsTo,
-  })),
+  graphql(SAVE_ADDRESS, { name: 'saveAddress' }),
   withContentWrapper(headerConfig),
-  withRuntimeContext
+  withRouter
 )
 
 export default enhance(AddressCreate)