+++ react/typings/graphql/getStoreConfigs.gql.d.ts
@@ -1,6 +1,0 @@
-declare module '*/getStoreConfigs.gql' {
-  import { DocumentNode } from 'graphql'
-
-  const value: DocumentNode
-  export default value
-}