+++ react/typings/graphql/logisticsInfo.gql.d.ts
@@ -1,7 +1,0 @@
-declare module '*/logisticsInfo.gql' {
-  import { DocumentNode } from 'graphql'
-
-  const value: DocumentNode
-
-  export default value
-}