+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Passing the newly created addressId to the return url.
+
 ## [1.13.5] - 2020-04-16
 
 ### Fixed